fix mod_xml_ldap build on OS X 10.6
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@15931 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/xml_int/mod_xml_ldap/mod_xml_ldap.c
@@ -353,7 +353,7 @@ static switch_xml_t xml_ldap_search(const char *section, const char *tag_name, c
 			goto cleanup;
 		}
 	} else {
-		if (ldap_sasl_interactive_bind_s( ldap->ld, NULL, binding->defaults->mech, NULL, NULL,(unsigned)LDAP_SASL_SIMPLE, lutil_sasl_interact , binding->defaults ) != LDAP_SUCCESS) {
+		if (ldap_sasl_interactive_bind_s(ldap->ld, NULL, binding->defaults->mech, NULL, NULL, (unsigned)(intptr_t)LDAP_SASL_SIMPLE, lutil_sasl_interact, binding->defaults ) != LDAP_SUCCESS) {
 			switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_ERROR, "Unable to sasl_bind to ldap server %s as %s\n", binding->host, binding->defaults->authcid);
 			goto cleanup;
 		}
